Added periodic task with celery
