+++ django_lab/django_lab/__init__.py
@@ -1,5 +1,3 @@
-from __future__ import absolute_import
-
 from .celery import app as celery_app
 
 __all__ = ('celery_app',)
