+++ django_lab/django_lab/celery.py
@@ -13,7 +13,7 @@ app = Celery('django_lab')
 app.config_from_object('django.conf:settings', namespace='CELERY')
 
 # Load task modules from all registered Django app configs.
-app.autodiscover_tasks()
+app.autodiscover_tasks(['main'])
 
 
 @app.task(bind=True)
