+++ django_lab/django_lab/settings.py
@@ -175,7 +175,7 @@ CKEDITOR_UPLOAD_PATH = "uploads/"
 # Celery Configuration Options
 CELERY_TIMEZONE = "Europe/Moscow"
 CELERY_BROKER_URL = 'redis://localhost:6379'
-CELERY_RESULT_BACKEND = 'redis://localhost:6379'
+CELERY_RESULT_BACKEND = 'db+sqlite:///results.sqlite'
 CELERY_ACCEPT_CONTENT = ['application/json']
 CELERY_TASK_SERIALIZER = 'json'
 CELERY_RESULT_SERIALIZER = 'json'
